# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KepRoop API Documentation",
    "version": "1.0"
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    },
    {
      "CookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Prefix with 'Api-Key ' or 'Bearer kp_'"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Browser Session Cookie (Automatically used if logged in)"
      }
    }
  },
  "paths": {
...
    "/api/albums/{id}/invite": {
      "post": {
        "tags": [
          "Albums"
        ],
        "summary": "Create invite",
        "description": "Generate an invite link.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "viewer",
                      "editor"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite created"
          }
        }
      },
      "delete": {
        "tags": [
          "Albums"
        ],
        "summary": "Revoke invite",
        "description": "Revoke an existing invite link.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "viewer",
                      "editor"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite revoked"
          }
        }
      },
      "get": {
        "tags": [
          "Albums"
        ],
        "summary": "Get invite",
        "description": "Get the current active invite link for a specific role.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "viewer",
                "editor"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite details"
          }
        }
      }
    },
    "/api/albums/{id}/members/{memberId}": {
      "patch": {
        "tags": [
          "Albums"
        ],
        "summary": "Update member role",
        "description": "Update the role of a member.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "viewer",
                      "editor",
                      "owner"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Albums"
        ],
        "summary": "Remove member",
        "description": "Remove a member from the album or leave the album.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed"
          }
        }
      }
    },
...
  },
...
}